+++ Editor/Utils/SaintsFieldSetupWindow.cs
@@ -1,12 +1,8 @@
 using System;
 using System.Collections;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-#if SAINTSFIELD_NEWTONSOFT_JSON
-using Newtonsoft.Json;
-#endif
-using SaintsField.Editor.Linq;
+// #if SAINTSFIELD_NEWTONSOFT_JSON
+// using Newtonsoft.Json;
+// #endif
 using SaintsField.Playa;
 using SaintsField.Utils;
 using UnityEditor;
@@ -57,29 +53,30 @@ namespace SaintsField.Editor.Utils
             if (!config.GetSetupWindowPopOnce())
             {
                 Debug.Log($"setupWindowPopOnce false, pop setup window for SaintsField");
-                GetWindow<SaintsFieldSetupWindow>("SaintsField Setup").Show();
-                EditorUtility.SetDirty(config);
-                config.setupWindowPopOnce = true;
+                EditorApplication.delayCall += () =>
+                {
+                    GetWindow<SaintsFieldSetupWindow>("SaintsField Setup").Show();
+                };
             }
         }
 
         // ReSharper disable InconsistentNaming
-        private struct ScopedRegistries
-        {
-            public string name;
-            public string url;
-            public List<string> scopes;
-        }
+        // private struct ScopedRegistries
+        // {
+        //     public string name;
+        //     public string url;
+        //     public List<string> scopes;
+        // }
+        //
+        // private class ManifestBase
+        // {
+        //     public Dictionary<string, string> dependencies;
+        // }
 
-        private class ManifestBase
-        {
-            public Dictionary<string, string> dependencies;
-        }
-
-        private class Manifest: ManifestBase
-        {
-            public List<ScopedRegistries> scopedRegistries;
-        }
+        // private class Manifest: ManifestBase
+        // {
+        //     public List<ScopedRegistries> scopedRegistries;
+        // }
         // ReSharper enable InconsistentNaming
 
 #pragma warning disable CS0414 // Type or member is only assigned
@@ -87,30 +84,40 @@ namespace SaintsField.Editor.Utils
         private bool _loadingCodeAnalysis;
 #pragma warning restore CS0414 // Type or member is only assigned
         // private bool _loadingUnitySerialization;
-        private const string ManifestFile = "Packages/manifest.json";
-
-
-        private const string scopeUrl = "https://package.openupm.com";
-        private static readonly ICollection<string> scopeScopes = new List<string>
-        {
-            "org.nuget.microsoft.codeanalysis.analyzers",
-            "org.nuget.microsoft.codeanalysis.common",
-            "org.nuget.microsoft.codeanalysis.csharp",
-            "org.nuget.system.buffers",
-            "org.nuget.system.collections.immutable",
-            "org.nuget.system.memory",
-            "org.nuget.system.numerics.vectors",
-            "org.nuget.system.reflection.metadata",
-            "org.nuget.system.runtime.compilerservices.unsafe",
-            "org.nuget.system.text.encoding.codepages",
-            "org.nuget.system.threading.tasks.extensions",
-        };
+        // private const string ManifestFile = "Packages/manifest.json";
+        //
+        //
+        // private const string scopeUrl = "https://package.openupm.com";
+        // private static readonly ICollection<string> scopeScopes = new List<string>
+        // {
+        //     "org.nuget.microsoft.codeanalysis.analyzers",
+        //     "org.nuget.microsoft.codeanalysis.common",
+        //     "org.nuget.microsoft.codeanalysis.csharp",
+        //     "org.nuget.system.buffers",
+        //     "org.nuget.system.collections.immutable",
+        //     "org.nuget.system.memory",
+        //     "org.nuget.system.numerics.vectors",
+        //     "org.nuget.system.reflection.metadata",
+        //     "org.nuget.system.runtime.compilerservices.unsafe",
+        //     "org.nuget.system.text.encoding.codepages",
+        //     "org.nuget.system.threading.tasks.extensions",
+        // };
 
         public override void OnEditorEnable()
         {
             _loadingSaintsEditor = false;
             _loadingCodeAnalysis = false;
             // _loadingUnitySerialization = false;
+        }
+
+        public override void OnEditorDestroy()
+        {
+            SaintsFieldConfig config = SaintsFieldConfigUtil.Config;
+            if(config != null && !config.setupWindowPopOnce)
+            {
+                EditorUtility.SetDirty(config);
+                config.setupWindowPopOnce = true;
+            }
         }
 
         #region Saints Editor
@@ -137,6 +144,7 @@ namespace SaintsField.Editor.Utils
         [PlayaDisableIf(true)]
 #endif
         [Button("Enable")]
+        // ReSharper disable once UnusedMember.Local
         private void EnableSaintsEditor()
         {
             _loadingSaintsEditor = true;
@@ -148,6 +156,7 @@ namespace SaintsField.Editor.Utils
         [DisableIf(true)]
 #endif
         [Button("Disable")]
+        // ReSharper disable once UnusedMember.Local
         private void DisableSaintsEditor()
         {
             _loadingSaintsEditor = true;
@@ -158,298 +167,298 @@ namespace SaintsField.Editor.Utils
 
         #endregion
 
-        #region Json
-#if !SAINTSFIELD_NEWTONSOFT_JSON
-        [Ordered]
-        private (EMessageType, string) InstallJsonStatus = (EMessageType.None, "");
-
-        [Ordered]
-        [InfoBox("Package com.unity.nuget.newtonsoft-json not installed, auto installer can not work", EMessageType.Error)]
-        [InfoBox("$" + nameof(InstallJsonStatus))]
-        [Button("Install newtonsoft-json")]
-        private IEnumerator InstallJson()
-        {
-            AddRequest _addRequest = Client.Add("com.unity.nuget.newtonsoft-json");
-            int counter = 0;
-            bool wait = true;
-            while (wait)
-            {
-                counter = (counter + 1) % 4;
-                switch (_addRequest.Status)
-                {
-                    case StatusCode.InProgress:
-                        InstallJsonStatus = (EMessageType.Warning, $"Installing Newtonsoft.Json, please wait{new string('.', counter)}");
-                        break;
-                    case StatusCode.Success:
-                        InstallJsonStatus = (EMessageType.Info, "Newtonsoft.Json Installed");
-                        wait = false;
-                        break;
-                    case StatusCode.Failure:
-                        InstallJsonStatus = (EMessageType.Error, $"Newtonsoft.Json install failed: {_addRequest.Error.message}");
-                        wait = false;
-                        break;
-                    default:
-                        throw new ArgumentOutOfRangeException();
-                }
-                yield return null;
-            }
-        }
-#endif
-        #endregion
-
-        #region Code Analysis
-
-        [Separator(10)]
-#if !SAINTSFIELD_NEWTONSOFT_JSON
-        [LayoutDisableIf(true)]
-        // [InfoBox("Package com.unity.nuget.newtonsoft-json not installed, auto installer can not work", EMessageType.Error)]
-#endif
-        [Ordered]
-        [LayoutStart("Layout System", ELayout.TitleBox)]
-
-        [AboveText("<u>Code Analysis</u> allows layout system to function more preciously on field orders", 5, 5)]
-        [Separator(5)]
-        [InfoBox("Loading, please wait...", show: nameof(_loadingCodeAnalysis))]
-
-        private static bool CodeAnalysisInstalled() => AppDomain.CurrentDomain.GetAssemblies()
-            .Any(a => a.FullName.StartsWith("Microsoft.CodeAnalysis.CSharp,"));
-
-        private string CodeAnalysisInstallInfo() => "<u>Code Analysis</u> is " +
-                                                   (CodeAnalysisInstalled()? "<color=green>installed</color>": "<color=brown>not installed</color>")
-                                                   + " in this project";
-
-        [Ordered]
-        [AboveText("$" + nameof(CodeAnalysisInstallInfo), 5, 5)]
-        [LayoutStart("./Code Analysis Install Buttons", ELayout.Horizontal)]
-        [DisableIf(nameof(CodeAnalysisInstalled))]
-        [Button("Install")]
-        private void InstallCodeAnalysis()
-        {
-#if SAINTSFIELD_NEWTONSOFT_JSON
-            string content = File.ReadAllText(ManifestFile);
-            Manifest manifest = JsonConvert.DeserializeObject<Manifest>(content);
-            bool foundDependencies = false;
-            // ReSharper disable once ForeachCanBePartlyConvertedToQueryUsingAnotherGetEnumerator
-            foreach (KeyValuePair<string, string> manifestDependency in manifest.dependencies)
-            {
-                // Debug.Log($"dependencies: {manifestDependency.Key}={manifestDependency.Value}");
-                // ReSharper disable once InvertIf
-                if (manifestDependency.Key == "org.nuget.microsoft.codeanalysis.csharp")
-                {
-                    foundDependencies = true;
-                    Debug.Log($"Already found dependencies {manifestDependency.Key}");
-                }
-            }
-
-            if (!foundDependencies)
-            {
-                manifest.dependencies["org.nuget.microsoft.codeanalysis.csharp"] = "4.14.0";
-                Debug.Log($"Add dependencies org.nuget.microsoft.codeanalysis.csharp={manifest.dependencies["org.nuget.microsoft.codeanalysis.csharp"]}");
-            }
-
-            const string scopeName = "package.openupm.com";
-
-            if (manifest.scopedRegistries == null)
-            {
-                manifest = new Manifest
-                {
-                    dependencies = manifest.dependencies,
-                    scopedRegistries = new List<ScopedRegistries>
-                    {
-                        new ScopedRegistries
-                        {
-                            name = scopeName,
-                            url = scopeUrl,
-                            scopes = scopeScopes.ToList(),
-                        },
-                    },
-                };
-            }
-            else
-            {
-                int foundScopedRegistriesIndex = -1;
-                ScopedRegistries foundScopedRegistriesValue = default;
-                foreach ((ScopedRegistries scopedRegistry, int index) in manifest.scopedRegistries.WithIndex())
-                {
-                    if (scopedRegistry.name == scopeName)
-                    {
-                        foundScopedRegistriesIndex = index;
-                        foundScopedRegistriesValue = scopedRegistry;
-                    }
-                }
-
-                if (foundScopedRegistriesIndex == -1)
-                {
-                    ScopedRegistries newCreated = new ScopedRegistries
-                    {
-                        name = scopeName,
-                        url = scopeUrl,
-                        scopes = scopeScopes.ToList(),
-                    };
-                    Debug.Log($"ScopedRegistries not found, add {newCreated.name}, {newCreated.url}, {string.Join(":", newCreated.scopes)}");
-                    manifest.scopedRegistries.Add(newCreated);
-                }
-                else
-                {
-                    ScopedRegistries oldUpdate = new ScopedRegistries
-                    {
-                        name = scopeName,
-                        url = scopeUrl,
-                        scopes = MergeList(foundScopedRegistriesValue.scopes, scopeScopes),
-                    };
-                    Debug.Log($"ScopedRegistries updated at {foundScopedRegistriesIndex} with result {oldUpdate.name}, {oldUpdate.url}, {string.Join(":", oldUpdate.scopes)}");
-                    manifest.scopedRegistries[foundScopedRegistriesIndex] = oldUpdate;
-                }
-            }
-
-            string jsonResult = JsonConvert.SerializeObject(manifest, Formatting.Indented);
-            Debug.Log(jsonResult);
-            _loadingCodeAnalysis = true;
-            File.WriteAllText(ManifestFile, jsonResult + "\n");
-#endif
-        }
-
-        private static List<string> MergeList(List<string> scopes, ICollection<string> toAdd)
-        {
-            List<string> result = new List<string>(scopes);
-            // ReSharper disable once ForeachCanBePartlyConvertedToQueryUsingAnotherGetEnumerator
-            foreach (string each in toAdd)
-            {
-                if (!result.Contains(each))
-                {
-                    Debug.Log($"add {each}");
-                    result.Add(each);
-                }
-            }
-
-            return result;
-        }
-
-        private readonly struct ProcessScopedRegistry
-        {
-            public readonly bool Delete;
-            public readonly int Index;
-            public readonly ScopedRegistries CleanedScopedRegistries;
-
-            public ProcessScopedRegistry(bool delete, int index, ScopedRegistries cleanedScopedRegistries)
-            {
-                Delete = delete;
-                Index = index;
-                CleanedScopedRegistries = cleanedScopedRegistries;
-            }
-        }
-
-        [Ordered]
-        [Button("Uninstall")]
-        [EnableIf(nameof(CodeAnalysisInstalled))]
-        private void UninstallCodeAnalysis()
-        {
-#if SAINTSFIELD_NEWTONSOFT_JSON
-            string content = File.ReadAllText(ManifestFile);
-            Manifest manifest = JsonConvert.DeserializeObject<Manifest>(content);
-
-            manifest.dependencies.Remove("org.nuget.microsoft.codeanalysis.csharp");
-
-            List<ProcessScopedRegistry> results = new List<ProcessScopedRegistry>();
-
-            ManifestBase result = manifest;
-            if (manifest.scopedRegistries != null)
-            {
-                foreach ((ScopedRegistries scopedRegistry, int index) in manifest.scopedRegistries.WithIndex().Reverse())
-                {
-                    if (scopedRegistry.url == scopeUrl)
-                    {
-                        var oldScopes = scopedRegistry.scopes;
-                        var newScopes = oldScopes.Except(scopeScopes).ToList();
-                        if (newScopes.Count == 0)
-                        {
-                            Debug.Log($"Delete {scopedRegistry.name}({scopedRegistry.url}) at {index}");
-                            results.Add(new ProcessScopedRegistry(true, index, default));
-                        }
-                        else
-                        {
-                            Debug.Log($"Update {scopedRegistry.name}({scopedRegistry.url}) at {index} with {string.Join(":", newScopes)}");
-                            results.Add(new ProcessScopedRegistry(false, index, new ScopedRegistries
-                            {
-                                name = scopedRegistry.name,
-                                url = scopedRegistry.url,
-                                scopes = newScopes,
-                            }));
-                        }
-                    }
-                }
-
-                foreach (ProcessScopedRegistry processScopedRegisty in results)
-                {
-                    if (processScopedRegisty.Delete)
-                    {
-                        manifest.scopedRegistries.RemoveAt(processScopedRegisty.Index);
-                    }
-                    else
-                    {
-                        manifest.scopedRegistries[processScopedRegisty.Index] = processScopedRegisty.CleanedScopedRegistries;
-                    }
-                }
-
-                if (manifest.scopedRegistries.Count == 0)
-                {
-                    Debug.Log("No scopedRegistries found, delete");
-                    result = new ManifestBase
-                    {
-                        dependencies = manifest.dependencies,
-                    };
-                }
-            }
-
-            string jsonResult = JsonConvert.SerializeObject(result, Formatting.Indented);
-            Debug.Log(jsonResult);
-            _loadingCodeAnalysis = true;
-            File.WriteAllText(ManifestFile, jsonResult + "\n");
-            SaintsMenu.RemoveCompileDefine(SaintsMenu.SAINTSFIELD_CODE_ANALYSIS);
-#endif
-        }
-
-        [LayoutEnd(".")]
-        [AboveText(
-            "<u>Code Analysis</u> is " +
-#if SAINTSFIELD_CODE_ANALYSIS
-            "<color=green>enabled</color>"
-#else
-            "<color=brown>not enabled</color>"
-#endif
-            + " in this project", 5, 5)]
-        [Ordered]
-#if SAINTSFIELD_CODE_ANALYSIS
-        [DisableIf(true)]
-#endif
-        [Button("Force Enable")]
-        private void EnableCodeAnalysis()
-        {
-            bool codeAnalysisFound = CodeAnalysisInstalled();
-            if (codeAnalysisFound)
-            {
-                SaintsMenu.AddCompileDefine(SaintsMenu.SAINTSFIELD_CODE_ANALYSIS);
-                _loadingCodeAnalysis = true;
-                return;
-            }
-
-            if (EditorUtility.DisplayDialog("Code Analysis Not Found",
-                    "Microsoft.CodeAnalysis.CSharp not found in your project. This will break your script compilation if it's not installed at all.\n" +
-                    "If you believe this is a detection mistake, you can enable it anyway",
-                    "Enable Anyway",
-                    "How To Install?"))
-            {
-                SaintsMenu.AddCompileDefine(SaintsMenu.SAINTSFIELD_CODE_ANALYSIS);
-                _loadingCodeAnalysis = true;
-            }
-            else
-            {
-                Application.OpenURL("https://github.com/TylerTemp/SaintsField/?tab=readme-ov-file#setup");
-            }
-        }
-
-        [LayoutEnd]
-        #endregion
+//         #region Json
+// #if !SAINTSFIELD_NEWTONSOFT_JSON
+//         [Ordered]
+//         private (EMessageType, string) InstallJsonStatus = (EMessageType.None, "");
+//
+//         [Ordered]
+//         [InfoBox("Package com.unity.nuget.newtonsoft-json not installed, auto installer can not work", EMessageType.Error)]
+//         [InfoBox("$" + nameof(InstallJsonStatus))]
+//         [Button("Install newtonsoft-json")]
+//         private IEnumerator InstallJson()
+//         {
+//             AddRequest _addRequest = Client.Add("com.unity.nuget.newtonsoft-json");
+//             int counter = 0;
+//             bool wait = true;
+//             while (wait)
+//             {
+//                 counter = (counter + 1) % 4;
+//                 switch (_addRequest.Status)
+//                 {
+//                     case StatusCode.InProgress:
+//                         InstallJsonStatus = (EMessageType.Warning, $"Installing Newtonsoft.Json, please wait{new string('.', counter)}");
+//                         break;
+//                     case StatusCode.Success:
+//                         InstallJsonStatus = (EMessageType.Info, "Newtonsoft.Json Installed");
+//                         wait = false;
+//                         break;
+//                     case StatusCode.Failure:
+//                         InstallJsonStatus = (EMessageType.Error, $"Newtonsoft.Json install failed: {_addRequest.Error.message}");
+//                         wait = false;
+//                         break;
+//                     default:
+//                         throw new ArgumentOutOfRangeException();
+//                 }
+//                 yield return null;
+//             }
+//         }
+// #endif
+//         #endregion
+//
+//         #region Code Analysis
+//
+//         [Separator(10)]
+// #if !SAINTSFIELD_NEWTONSOFT_JSON
+//         [LayoutDisableIf(true)]
+//         // [InfoBox("Package com.unity.nuget.newtonsoft-json not installed, auto installer can not work", EMessageType.Error)]
+// #endif
+//         [Ordered]
+//         [LayoutStart("Layout System", ELayout.TitleBox)]
+//
+//         [AboveText("<u>Code Analysis</u> allows layout system to function more preciously on field orders", 5, 5)]
+//         [Separator(5)]
+//         [InfoBox("Loading, please wait...", show: nameof(_loadingCodeAnalysis))]
+//
+//         private static bool CodeAnalysisInstalled() => AppDomain.CurrentDomain.GetAssemblies()
+//             .Any(a => a.FullName.StartsWith("Microsoft.CodeAnalysis.CSharp,"));
+//
+//         private string CodeAnalysisInstallInfo() => "<u>Code Analysis</u> is " +
+//                                                    (CodeAnalysisInstalled()? "<color=green>installed</color>": "<color=brown>not installed</color>")
+//                                                    + " in this project";
+//
+//         [Ordered]
+//         [AboveText("$" + nameof(CodeAnalysisInstallInfo), 5, 5)]
+//         [LayoutStart("./Code Analysis Install Buttons", ELayout.Horizontal)]
+//         [DisableIf(nameof(CodeAnalysisInstalled))]
+//         [Button("Install")]
+//         private void InstallCodeAnalysis()
+//         {
+// #if SAINTSFIELD_NEWTONSOFT_JSON
+//             string content = File.ReadAllText(ManifestFile);
+//             Manifest manifest = JsonConvert.DeserializeObject<Manifest>(content);
+//             bool foundDependencies = false;
+//             // ReSharper disable once ForeachCanBePartlyConvertedToQueryUsingAnotherGetEnumerator
+//             foreach (KeyValuePair<string, string> manifestDependency in manifest.dependencies)
+//             {
+//                 // Debug.Log($"dependencies: {manifestDependency.Key}={manifestDependency.Value}");
+//                 // ReSharper disable once InvertIf
+//                 if (manifestDependency.Key == "org.nuget.microsoft.codeanalysis.csharp")
+//                 {
+//                     foundDependencies = true;
+//                     Debug.Log($"Already found dependencies {manifestDependency.Key}");
+//                 }
+//             }
+//
+//             if (!foundDependencies)
+//             {
+//                 manifest.dependencies["org.nuget.microsoft.codeanalysis.csharp"] = "4.14.0";
+//                 Debug.Log($"Add dependencies org.nuget.microsoft.codeanalysis.csharp={manifest.dependencies["org.nuget.microsoft.codeanalysis.csharp"]}");
+//             }
+//
+//             const string scopeName = "package.openupm.com";
+//
+//             if (manifest.scopedRegistries == null)
+//             {
+//                 manifest = new Manifest
+//                 {
+//                     dependencies = manifest.dependencies,
+//                     scopedRegistries = new List<ScopedRegistries>
+//                     {
+//                         new ScopedRegistries
+//                         {
+//                             name = scopeName,
+//                             url = scopeUrl,
+//                             scopes = scopeScopes.ToList(),
+//                         },
+//                     },
+//                 };
+//             }
+//             else
+//             {
+//                 int foundScopedRegistriesIndex = -1;
+//                 ScopedRegistries foundScopedRegistriesValue = default;
+//                 foreach ((ScopedRegistries scopedRegistry, int index) in manifest.scopedRegistries.WithIndex())
+//                 {
+//                     if (scopedRegistry.name == scopeName)
+//                     {
+//                         foundScopedRegistriesIndex = index;
+//                         foundScopedRegistriesValue = scopedRegistry;
+//                     }
+//                 }
+//
+//                 if (foundScopedRegistriesIndex == -1)
+//                 {
+//                     ScopedRegistries newCreated = new ScopedRegistries
+//                     {
+//                         name = scopeName,
+//                         url = scopeUrl,
+//                         scopes = scopeScopes.ToList(),
+//                     };
+//                     Debug.Log($"ScopedRegistries not found, add {newCreated.name}, {newCreated.url}, {string.Join(":", newCreated.scopes)}");
+//                     manifest.scopedRegistries.Add(newCreated);
+//                 }
+//                 else
+//                 {
+//                     ScopedRegistries oldUpdate = new ScopedRegistries
+//                     {
+//                         name = scopeName,
+//                         url = scopeUrl,
+//                         scopes = MergeList(foundScopedRegistriesValue.scopes, scopeScopes),
+//                     };
+//                     Debug.Log($"ScopedRegistries updated at {foundScopedRegistriesIndex} with result {oldUpdate.name}, {oldUpdate.url}, {string.Join(":", oldUpdate.scopes)}");
+//                     manifest.scopedRegistries[foundScopedRegistriesIndex] = oldUpdate;
+//                 }
+//             }
+//
+//             string jsonResult = JsonConvert.SerializeObject(manifest, Formatting.Indented);
+//             Debug.Log(jsonResult);
+//             _loadingCodeAnalysis = true;
+//             File.WriteAllText(ManifestFile, jsonResult + "\n");
+// #endif
+//         }
+//
+//         private static List<string> MergeList(List<string> scopes, ICollection<string> toAdd)
+//         {
+//             List<string> result = new List<string>(scopes);
+//             // ReSharper disable once ForeachCanBePartlyConvertedToQueryUsingAnotherGetEnumerator
+//             foreach (string each in toAdd)
+//             {
+//                 if (!result.Contains(each))
+//                 {
+//                     Debug.Log($"add {each}");
+//                     result.Add(each);
+//                 }
+//             }
+//
+//             return result;
+//         }
+//
+//         private readonly struct ProcessScopedRegistry
+//         {
+//             public readonly bool Delete;
+//             public readonly int Index;
+//             public readonly ScopedRegistries CleanedScopedRegistries;
+//
+//             public ProcessScopedRegistry(bool delete, int index, ScopedRegistries cleanedScopedRegistries)
+//             {
+//                 Delete = delete;
+//                 Index = index;
+//                 CleanedScopedRegistries = cleanedScopedRegistries;
+//             }
+//         }
+//
+//         [Ordered]
+//         [Button("Uninstall")]
+//         [EnableIf(nameof(CodeAnalysisInstalled))]
+//         private void UninstallCodeAnalysis()
+//         {
+// #if SAINTSFIELD_NEWTONSOFT_JSON
+//             string content = File.ReadAllText(ManifestFile);
+//             Manifest manifest = JsonConvert.DeserializeObject<Manifest>(content);
+//
+//             manifest.dependencies.Remove("org.nuget.microsoft.codeanalysis.csharp");
+//
+//             List<ProcessScopedRegistry> results = new List<ProcessScopedRegistry>();
+//
+//             ManifestBase result = manifest;
+//             if (manifest.scopedRegistries != null)
+//             {
+//                 foreach ((ScopedRegistries scopedRegistry, int index) in manifest.scopedRegistries.WithIndex().Reverse())
+//                 {
+//                     if (scopedRegistry.url == scopeUrl)
+//                     {
+//                         var oldScopes = scopedRegistry.scopes;
+//                         var newScopes = oldScopes.Except(scopeScopes).ToList();
+//                         if (newScopes.Count == 0)
+//                         {
+//                             Debug.Log($"Delete {scopedRegistry.name}({scopedRegistry.url}) at {index}");
+//                             results.Add(new ProcessScopedRegistry(true, index, default));
+//                         }
+//                         else
+//                         {
+//                             Debug.Log($"Update {scopedRegistry.name}({scopedRegistry.url}) at {index} with {string.Join(":", newScopes)}");
+//                             results.Add(new ProcessScopedRegistry(false, index, new ScopedRegistries
+//                             {
+//                                 name = scopedRegistry.name,
+//                                 url = scopedRegistry.url,
+//                                 scopes = newScopes,
+//                             }));
+//                         }
+//                     }
+//                 }
+//
+//                 foreach (ProcessScopedRegistry processScopedRegisty in results)
+//                 {
+//                     if (processScopedRegisty.Delete)
+//                     {
+//                         manifest.scopedRegistries.RemoveAt(processScopedRegisty.Index);
+//                     }
+//                     else
+//                     {
+//                         manifest.scopedRegistries[processScopedRegisty.Index] = processScopedRegisty.CleanedScopedRegistries;
+//                     }
+//                 }
+//
+//                 if (manifest.scopedRegistries.Count == 0)
+//                 {
+//                     Debug.Log("No scopedRegistries found, delete");
+//                     result = new ManifestBase
+//                     {
+//                         dependencies = manifest.dependencies,
+//                     };
+//                 }
+//             }
+//
+//             string jsonResult = JsonConvert.SerializeObject(result, Formatting.Indented);
+//             Debug.Log(jsonResult);
+//             _loadingCodeAnalysis = true;
+//             File.WriteAllText(ManifestFile, jsonResult + "\n");
+//             SaintsMenu.RemoveCompileDefine(SaintsMenu.SAINTSFIELD_CODE_ANALYSIS);
+// #endif
+//         }
+//
+//         [LayoutEnd(".")]
+//         [AboveText(
+//             "<u>Code Analysis</u> is " +
+// #if SAINTSFIELD_CODE_ANALYSIS
+//             "<color=green>enabled</color>"
+// #else
+//             "<color=brown>not enabled</color>"
+// #endif
+//             + " in this project", 5, 5)]
+//         [Ordered]
+// #if SAINTSFIELD_CODE_ANALYSIS
+//         [DisableIf(true)]
+// #endif
+//         [Button("Force Enable")]
+//         private void EnableCodeAnalysis()
+//         {
+//             bool codeAnalysisFound = CodeAnalysisInstalled();
+//             if (codeAnalysisFound)
+//             {
+//                 SaintsMenu.AddCompileDefine(SaintsMenu.SAINTSFIELD_CODE_ANALYSIS);
+//                 _loadingCodeAnalysis = true;
+//                 return;
+//             }
+//
+//             if (EditorUtility.DisplayDialog("Code Analysis Not Found",
+//                     "Microsoft.CodeAnalysis.CSharp not found in your project. This will break your script compilation if it's not installed at all.\n" +
+//                     "If you believe this is a detection mistake, you can enable it anyway",
+//                     "Enable Anyway",
+//                     "How To Install?"))
+//             {
+//                 SaintsMenu.AddCompileDefine(SaintsMenu.SAINTSFIELD_CODE_ANALYSIS);
+//                 _loadingCodeAnalysis = true;
+//             }
+//             else
+//             {
+//                 Application.OpenURL("https://github.com/TylerTemp/SaintsField/?tab=readme-ov-file#setup");
+//             }
+//         }
+//
+//         [LayoutEnd]
+//         #endregion
 
         #region Unity Serialization
 
@@ -480,6 +489,7 @@ namespace SaintsField.Editor.Utils
         [DisableIf(true)]
 #endif
         [Button("Install")]
+        // ReSharper disable once UnusedMember.Local
         private IEnumerator InstallUnitySerialization()
         {
             const string packageName = "com.unity.serialization";
@@ -537,6 +547,7 @@ namespace SaintsField.Editor.Utils
 #if !SAINTSFIELD_SERIALIZATION
         [PlayaDisableIf(true)]
 #endif
+        // ReSharper disable once UnusedMember.Local
         private IEnumerator UninstallUnitySerialization()
         {
             const string packageName = "com.unity.serialization";